remove use of deprecated defaultPartition
