+++ hapi-fhir-jpaserver-searchparam/src/main/java/ca/uhn/fhir/jpa/partition/IRequestPartitionHelperSvc.java
@@ -21,6 +21,7 @@ package ca.uhn.fhir.jpa.partition;
 
 import ca.uhn.fhir.interceptor.model.ReadPartitionIdRequestDetails;
 import ca.uhn.fhir.interceptor.model.RequestPartitionId;
+import ca.uhn.fhir.interceptor.model.IDefaultPartitionSettings;
 import ca.uhn.fhir.jpa.searchparam.SearchParameterMap;
 import ca.uhn.fhir.rest.api.server.RequestDetails;
 import jakarta.annotation.Nonnull;
@@ -30,7 +31,7 @@ import org.hl7.fhir.instance.model.api.IIdType;
 
 import java.util.Set;
 
-public interface IRequestPartitionHelperSvc {
+public interface IRequestPartitionHelperSvc extends IDefaultPartitionSettings {
 
 	@Nonnull
 	RequestPartitionId determineReadPartitionForRequest(
@@ -182,37 +183,4 @@ public interface IRequestPartitionHelperSvc {
 	 * @return - A {@link RequestPartitionId} with a normalized list of partition ids and partition names.
 	 */
 	RequestPartitionId validateAndNormalizePartitionNames(RequestPartitionId theRequestPartitionId);
-
-	/**
-	 * This method returns the default partition ID. Implementers of this interface should overwrite this method to provide
-	 * a default partition ID that is different than the default value of null.
-	 *
-	 * @return the default partition ID
-	 */
-	@Nullable
-	default Integer getDefaultPartitionId() {
-		return null;
-	}
-
-	/**
-	 * Test whether <code>theRequestPartitionId</code> is only targeting the default partition where the ID of the default
-	 * partition is provided by {@link #getDefaultPartitionId()}.
-	 *
-	 * @param theRequestPartitionId to perform the evaluation upon.
-	 * @return true if the <code>theRequestPartitionId</code> is for the default partition only.
-	 */
-	default boolean isDefaultPartition(@Nonnull RequestPartitionId theRequestPartitionId) {
-		return theRequestPartitionId.isDefaultPartition(getDefaultPartitionId());
-	}
-
-	/**
-	 * Test whether <code>theRequestPartitionId</code> has one of its targeted partitions matching the default partition
-	 * where the ID of the default partition is provided by {@link #getDefaultPartitionId()}.
-	 *
-	 * @param theRequestPartitionId to perform the evaluation upon.
-	 * @return true if the <code>theRequestPartitionId</code> is targeting the default partition.
-	 */
-	default boolean hasDefaultPartitionId(@Nonnull RequestPartitionId theRequestPartitionId) {
-		return theRequestPartitionId.hasDefaultPartitionId(getDefaultPartitionId());
-	}
 }
